- Fix security risks - Refactor code
