+++ app/ussd/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-api_v2 = Blueprint('api_v2', __name__)
+ussd = Blueprint('ussd', __name__)
 
 from . import views, decorators
