+++ requirements.txt
@@ -1,7 +1,7 @@
 AfricastalkingGateway==1.9
 psycopg2==2.7.7
 celery==4.1.0
-Flask==0.12.2
+flask>=0.12.3
 Flask-Bootstrap==3.3.7.1
 Flask-Login==0.4.0
 Flask-Migrate==2.1.1
@@ -12,3 +12,4 @@ gunicorn==19.7.1
 blinker==1.4
 Flask_Redis==0.3.0
 python-dotenv
+pyaml
